need to fix constructors issue

+++ EX03 Running/B21 Ex03 Eithan 204311757 Maor 204709950/Ex03.GarageLogic/Object classes/GasCar.cs
@@ -16,7 +16,7 @@ namespace Ex03.GarageLogic
         /// <exception cref="ArgumentException"></exception>
         public GasCar(string i_Model, string i_PlateID, Color i_Color, NumOfDoors i_NumOfDoors, GasType i_GasType, 
             float i_FuelLeft, float i_MaxFuel, string[] i_WheelsManufacturers, float[] i_WheelsCurrentAirPressures) :
-            base(i_Model, i_PlateID, (i_FuelLeft / i_MaxFuel) * 100, i_Color, i_NumOfDoors,
+            base(i_Model, i_PlateID, (float)(i_FuelLeft / i_MaxFuel * 100), i_Color, i_NumOfDoors,
                 i_WheelsManufacturers, i_WheelsCurrentAirPressures)
         {
             m_GasType = i_GasType;
